frontend/requirements.txt
attrs==22.2.0
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==39.0.0
Cython==0.29.32
docutils==0.19
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
jaraco.classes==3.2.3
jeepney==0.8.0
kaki==0.1.5
Kivy @ git+https://github.com/kivy/kivy.git@023bd79b90f9831b45bb8eb449346648aa5fe5f8
Kivy-Garden==0.1.5
kivymd==1.1.1
monotonic==1.6
more-itertools==9.0.0
Pillow==9.3.0
pycparser==2.21
pycryptodome==3.16.0
Pygments==2.13.0
requests==2.28.1
SecretStorage==3.3.3
six==1.16.0
Twisted==22.10.0
typing-extensions==4.4.0
urllib3==1.26.12
watchdog==2.1.9
zipp==3.11.0
zope.interface==5.5.2
